Move main project resolvers to factories
