+++ test/Npgsql.Tests/Types/LegacyDateTimeTests.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Data;
 using System.Threading.Tasks;
-using Npgsql.Internal.Resolvers;
 using NpgsqlTypes;
 using NUnit.Framework;
 using static Npgsql.Util.Statics;
